Changed the buttons
All buttons now have an animator attached to them - but it needs to be
told thru code when to act!

I'd suggest writing a script checking when the player is hovering their
mouse over the button and calling the animator when that is the case.
Same thing with clicking buttons.

The animator's variables are "onhover" and "click"

I can probably do it myself when I'm less busy :'D